egui replaced with built-in ui

+++ src/settings/mod.rs
@@ -2,12 +2,12 @@ pub mod audio;
 pub mod window;
 
 pub struct AudioSettings {
-	pub music_volume: u8,
-	pub sound_volume: u8,
+	pub music_volume: f32,
+	pub sound_volume: f32,
 }
 
 pub struct WindowSettings {
 	pub fullscreen: bool,
 	pub resolution_scale: f32,
-	pub resolution: u8,
+	pub resolution: usize,
 }
